refactor: consolidate output duplication and merge collect files
Two simplifications to reduce code complexity:

1. Consolidate background/foreground output duplication in handlers.rs:
   - Create RemovalDisplayInfo struct to capture removal outcome
   - Unify message formatting logic between modes
   - Fix warning order (foreground: after progress, background: before decision)

2. Merge collect_progressive_impl.rs into collect.rs:
   - Eliminate artificial file split with circular dependencies
   - Update visibility from pub(super) to pub(crate) for merged types
   - Net reduction of ~180 lines
